+++ benchmark/benchmark_merge_path.hpp
@@ -14,7 +14,7 @@
 #include "utils/benchmark_time.h"
 
 struct MergePathSpMV : CsrSpMV {
-  void csr_spmv_impl(int trans, const int alpha, const int beta, const csr_desc<int, double> h_csr_desc,
+  void csr_spmv_impl(int trans, const double alpha, const double beta, const csr_desc<int, double> h_csr_desc,
                      const csr_desc<int, double> d_csr_desc, const double *x, double *y, BenchmarkTime *bmt) {
     merge_path_spmv(trans, alpha, beta, h_csr_desc, d_csr_desc, x, y, bmt);
   }
@@ -22,7 +22,7 @@ struct MergePathSpMV : CsrSpMV {
 };
 
 struct MergePathSingleBlockUpdateSpMV : CsrSpMV {
-  void csr_spmv_impl(int trans, const int alpha, const int beta, const csr_desc<int, double> h_csr_desc,
+  void csr_spmv_impl(int trans, const double alpha, const double beta, const csr_desc<int, double> h_csr_desc,
                      const csr_desc<int, double> d_csr_desc, const double *x, double *y, BenchmarkTime *bmt) {
     merge_path_spmv<Binary, SingleBlock>(trans, alpha, beta, h_csr_desc, d_csr_desc, x, y, bmt);
   }
@@ -30,7 +30,7 @@ struct MergePathSingleBlockUpdateSpMV : CsrSpMV {
 };
 
 struct MergePathLookBackUpdateSpMV : CsrSpMV {
-  void csr_spmv_impl(int trans, const int alpha, const int beta, const csr_desc<int, double> h_csr_desc,
+  void csr_spmv_impl(int trans, const double alpha, const double beta, const csr_desc<int, double> h_csr_desc,
                      const csr_desc<int, double> d_csr_desc, const double *x, double *y, BenchmarkTime *bmt) {
     merge_path_spmv<Binary, LookBack>(trans, alpha, beta, h_csr_desc, d_csr_desc, x, y, bmt);
   }
