+++ benchmark/csr_spmv.hpp
@@ -38,7 +38,7 @@ struct CsrSpMV {
    * @param y device vector y.
    * @param bmt spmv elapsed time
    */
-  virtual void csr_spmv_impl(int trans, const int alpha, const int beta,
+  virtual void csr_spmv_impl(int trans, const double alpha, const double beta,
                              const csr_desc<int, double> h_csr_desc, const csr_desc<int, double> d_csr_desc,
                              const double *x, double *y, BenchmarkTime *bmt) = 0;
